confluent/stage-gcp/cloud-functions/context_client/requirements.txt
google-cloud-aiplatform
requests